fix some tests issues for login api routes, delete all unused variables, move auth config options out from route file
