+++ nextjs-metrics-viewer/jest.config.ts
@@ -1,7 +1,10 @@
 import type { Config } from "jest";
 import nextJest from "next/jest.js";
+import { pathsToModuleNameMapper } from "ts-jest";
+import { compilerOptions } from "./tsconfig.json";
 
 const createJestConfig = nextJest({
+  // Provide the path to your Next.js app to load next.config.js and .env files in your test environment
   dir: "./",
 });
 
@@ -9,12 +12,12 @@ const createJestConfig = nextJest({
 const config: Config = {
   coverageProvider: "v8",
   testEnvironment: "jsdom",
-  moduleFileExtensions: ["ts", "js", "json"],
-  setupFiles: ["<rootDir>/setup-jest.ts"],
-  moduleNameMapper: {
-    "^@/components/(.*)$": "<rootDir>/components/$1",
-    "^@/(.*)$": "<rootDir>/$1",
-  },
+  // Add more setup options before each test is run
+  setupFiles: ["<rootDir>/jest.setup.ts"],
+  moduleNameMapper: pathsToModuleNameMapper(compilerOptions.paths, {
+    prefix: "<rootDir>/",
+  }),
 };
 
+// createJestConfig is exported this way to ensure that next/jest can load the Next.js config which is async
 export default createJestConfig(config);
