+++ nextjs-metrics-viewer/jest.setup.ts
@@ -1,0 +1,5 @@
+import dotenv from "dotenv";
+
+dotenv.config({ path: ".env.test.local" });
+
+jest.spyOn(console, "error").mockImplementation(() => {});
