+++ nextjs-metrics-viewer/package.json
@@ -63,6 +63,7 @@
     "jest-environment-jsdom": "^29.7.0",
     "jest-mock-extended": "4.0.0-beta1",
     "node-fetch": "^3.3.2",
+    "node-mocks-http": "^1.16.1",
     "postcss": "^8",
     "prettier": "^3.2.5",
     "prettier-plugin-tailwindcss": "^0.5.12",
