+++ nextjs-metrics-viewer/test/routes/auth/login.test.ts
@@ -1,66 +1,82 @@
-import { StatusCodes } from "http-status-codes";
+/**
+ * @jest-environment node
+ */
+import { createMocks, RequestMethod } from "node-mocks-http";
+import type { NextApiRequest } from "next";
 import { POST } from "@/app/api/auth/[...nextauth]/route";
-import { DatabaseConnection } from "@/lib/db";
-import { users } from "@/lib/db/schemas/user";
+import { AppError } from "@/lib/errors";
 import { UserFactory } from "@/test/factories/users";
-import UserUtilities from "@/domain/users/utilities";
-import { eq } from "drizzle-orm";
 
-jest.mock("next/server", () => ({
-  NextResponse: {
-    json: jest.fn((data, { status }) => ({
-      json: () => data,
-      status,
-    })),
+const mockUser = UserFactory.validUser().build();
+
+jest.mock("@/lib/db", () => ({
+  DatabaseConnection: {
+    getInstance: jest.fn(() => {
+      const instance = {};
+      Object.defineProperty(instance, "db", {
+        get: jest.fn(() => ({
+          query: jest.fn().mockResolvedValue([]),
+        })),
+      });
+      return instance;
+    }),
   },
 }));
 
-describe("POST /app/api/auth/[...nextauth]/route", () => {
-  const mockUser = UserFactory.validUser().build();
+jest.mock("@/app/api/auth/[...nextauth]/services/login", () => ({
+  LoginService: jest.fn().mockImplementation(() => ({
+    perform: jest.fn().mockResolvedValue({
+      user: mockUser,
+      token: "mockToken",
+    }),
+  })),
+}));
 
-  beforeAll(async () => {
-    const db = DatabaseConnection.getInstance().db;
-    await db.insert(users).values({
-      ...mockUser,
-      password: await UserUtilities.encryptPassword(mockUser.password),
+describe("POST /api/auth/login", () => {
+  function mockRequestResponse(method: RequestMethod = "POST") {
+    const { req, res }: { req: NextApiRequest; res: any } = createMocks({
+      method,
     });
+    req.headers = {
+      "Content-Type": "application/json",
+    };
+    return { req, res };
+  }
+
+  it("should authenticate user with valid credentials", async () => {
+    const { req, res } = mockRequestResponse();
+    req.body = { email: mockUser.email, password: mockUser.password };
+
+    await POST(req, res);
+
+    expect(res._getStatusCode()).toBe(200);
   });
 
-  beforeEach(() => {
-    jest.clearAllMocks();
+  it("should return 400 if credentials are missing", async () => {
+    const { req, res } = mockRequestResponse();
+    req.body = {};
+
+    await POST(req, res);
+
+    expect(res._getStatusCode()).toBe(400);
   });
 
-  afterAll(async () => {
-    const db = DatabaseConnection.getInstance().db;
-    await db.delete(users).where(eq(users.email, mockUser.email));
-  });
+  it("should return 500 if an internal error occurs", async () => {
+    jest.spyOn(console, "error").mockImplementation(() => {});
 
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  it("should return 200 and user data on successful login", async () => {
-    const req = new Request("http://localhost/api/auth/login", {
-      method: "POST",
-      body: JSON.stringify({
-        email: mockUser.email,
-        password: mockUser.password,
-      }),
+    const mockLoginService = jest.requireMock(
+      "@/app/api/auth/[...nextauth]/services/login",
+    ).LoginService;
+    mockLoginService.mockImplementationOnce(() => {
+      throw new AppError("Internal Server Error", 500);
     });
 
-    const res = await POST(req);
+    const { req, res } = mockRequestResponse();
+    req.body = { email: mockUser.email, password: mockUser.password };
 
-    expect(res.status).toBe(StatusCodes.OK);
-  });
+    await POST(req, res);
 
-  it("should return 400 when email or password is missing", async () => {
-    const req = new Request("http://localhost/api/auth/login", {
-      method: "POST",
-      body: JSON.stringify({ email: mockUser.email }),
-    });
-
-    const res = await POST(req);
-
-    expect(res.status).toBe(StatusCodes.BAD_REQUEST);
+    expect(res._getStatusCode()).toBe(500);
+    jest.restoreAllMocks();
   });
 });
